Move client/mapview* to client/views with rename
