+++ client/citybar.cpp
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 1996-2020 Freeciv and Freeciv21 contributors. This file is
+ Copyright (c) 1996-2023 Freeciv and Freeciv21 contributors. This file is
  __    __          part of Freeciv21. Freeciv21 is free software: you can
 / \\..// \    redistribute it and/or modify it under the terms of the GNU
   ( oo )        General Public License  as published by the Free Software
@@ -29,7 +29,7 @@
 // client
 #include "client_main.h"
 #include "colors_common.h"
-#include "mapview_common.h"
+#include "views/view_map_common.h"
 
 #include "citybar.h"
