+++ client/hudwidget.cpp
@@ -31,17 +31,17 @@
 #include "calendar.h"
 #include "client_main.h"
 #include "goto.h"
-#include "mapview_common.h"
 #include "text.h"
+#include "views/view_map_common.h"
 // gui-qt
 #include "canvas.h"
 #include "fc_client.h"
 #include "fonts.h"
 #include "icons.h"
-#include "mapview.h"
 #include "page_game.h"
 #include "tileset/sprite.h"
 #include "top_bar.h"
+#include "views/view_map.h"
 #include "widgets/decorations.h"
 
 static QString popup_terrain_info(struct tile *ptile);
