+++ client/notifyreport.cpp
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 1996-2020 Freeciv21 and Freeciv contributors. This file is
+ Copyright (c) 1996-2023 Freeciv21 and Freeciv contributors. This file is
  part of Freeciv21. Freeciv21 is free software: you can redistribute it
  and/or modify it under the terms of the GNU  General Public License  as
  published by the Free Software Foundation, either version 3 of the
@@ -17,8 +17,8 @@
 // gui-qt
 #include "fc_client.h"
 #include "fonts.h"
-#include "mapview.h"
 #include "page_game.h"
+#include "views/view_map.h"
 
 /**
    Constructor for notify dialog
