+++ client/options.cpp
@@ -54,13 +54,13 @@
 #include "global_worklist.h"
 #include "governor.h"
 #include "mapctrl_common.h"
-#include "mapview_common.h"
 #include "music.h"
 #include "overview_common.h"
 #include "packhand_gen.h"
 #include "themes_common.h"
 #include "tileset/tilespec.h"
 #include "views/view_cities_data.h"
+#include "views/view_map_common.h"
 #include "views/view_nations_data.h"
 
 #include "options.h"
