+++ client/top_bar.cpp
@@ -34,12 +34,12 @@
 // gui-qt
 #include "fc_client.h"
 #include "fonts.h"
-#include "mapview.h"
 #include "page_game.h"
 #include "qtg_cxxside.h"
 #include "sciencedlg.h"
 #include "tileset/sprite.h"
 #include "top_bar.h"
+#include "views/view_map.h"
 
 /**
  * Constructor
