+++ client/views/view_map_common.h
