+++ client/views/view_map_geometry.h
